+++ src/item/Item.js
@@ -1588,6 +1588,15 @@ new function() { // Injection scope for various item event handlers
         return this._index;
     },
 
+    setIndex: function(index) {
+        var parent = this._parent,
+            children = parent && parent._children;
+        if (children) {
+            parent.insertChildren(index in children ? index : undefined,
+                                  [this]);
+        }
+    },
+
     equals: function(item) {
         // NOTE: We do not compare name and selected state.
         // TODO: Consider not comparing locked and visible also?
@@ -2447,6 +2456,9 @@ new function() { // Injection scope for hit-test functions shared with project
      * @option [options.embedImages=true] {Boolean} whether raster images should
      *     be embedded as base64 data inlined in the xlink:href attribute, or
      *     kept as a link to their external URL.
+     * @option [options.reduceAttributes=true] {Boolean} wether to only include
+     *     style attributes in the SVG output that differ from their parents,
+     *     or to always include them (much faster but leading to redundancies).
      *
      * @param {Object} [options] the export options
      * @return {SVGElement|String} the item converted to an SVG node or a