+++ src/svg/SvgImport.js
@@ -333,7 +333,7 @@ new function() {
                     break;
                 // Command come before the '(', values after
                 var parts = transform.split(/\(\s*/),
-                    command = parts[0],
+                    command = parts[0].trim(),
                     v = parts[1].split(/[\s,]+/g);
                 // Convert values to floats
                 for (var j = 0, m = v.length; j < m; j++)
@@ -373,7 +373,7 @@ new function() {
             color.setAlpha(parseFloat(value));
     }
 
-    // Create apply-functions for attributes, and merge in those for SVGStlyes.
+    // Create apply-functions for attributes, and merge in those for SVGStyles.
     // We need to define style attributes first, and merge in all others after,
     // since transform needs to be applied after fill color, as transformations
     // can affect gradient fills.